+++ c/ndsort.c
@@ -199,7 +199,7 @@ int main(int argc, char *argv[])
         printf ("# points: %d\n", size);
     }
 
-    int * rank = pareto_rank (points, dim, size);
+    int * rank = pareto_rank(points, size, dim);
 
     if (only_rank_flag) {
         fprint_rank (stdout, rank, size);